Multiple changes
* Consolidate Signer and Verifier interfaces in signerverifier.go
* Move verification tests to verify_test.go
* Rename interceptSigner
