+++ dsse/sign.go
@@ -8,45 +8,10 @@ import (
 	"context"
 	"encoding/base64"
 	"errors"
-	"fmt"
 )
-
-// ErrUnknownKey indicates that the implementation does not recognize the
-// key.
-var ErrUnknownKey = errors.New("unknown key")
-
-// ErrNoSignature indicates that an envelope did not contain any signatures.
-var ErrNoSignature = errors.New("no signature found")
 
 // ErrNoSigners indicates that no signer was provided.
 var ErrNoSigners = errors.New("no signers provided")
-
-/*
-PAE implementes the DSSE Pre-Authentic Encoding
-https://github.com/secure-systems-lab/dsse/blob/master/protocol.md#signature-definition
-*/
-func PAE(payloadType string, payload []byte) []byte {
-	return []byte(fmt.Sprintf("DSSEv1 %d %s %d %s",
-		len(payloadType), payloadType,
-		len(payload), payload))
-}
-
-/*
-Signer defines the interface for an abstract signing algorithm.
-The Signer interface is used to inject signature algorithm implementations
-into the EnevelopeSigner. This decoupling allows for any signing algorithm
-and key management system can be used.
-The full message is provided as the parameter. If the signature algorithm
-depends on hashing of the message prior to signature calculation, the
-implementor of this interface must perform such hashing.
-The function must return raw bytes representing the calculated signature
-using the current algorithm, and the key used (if applicable).
-For an example see EcdsaSigner in sign_test.go.
-*/
-type Signer interface {
-	Sign(ctx context.Context, data []byte) ([]byte, error)
-	KeyID() (string, error)
-}
 
 // EnvelopeSigner creates signed Envelopes.
 type EnvelopeSigner struct {
