+++ dsse/verify.go
@@ -9,17 +9,8 @@ import (
 	"golang.org/x/crypto/ssh"
 )
 
-/*
-Verifier verifies a complete message against a signature and key.
-If the message was hashed prior to signature generation, the verifier
-must perform the same steps.
-If KeyID returns successfully, only signature matching the key ID will be verified.
-*/
-type Verifier interface {
-	Verify(ctx context.Context, data, sig []byte) error
-	KeyID() (string, error)
-	Public() crypto.PublicKey
-}
+// ErrNoSignature indicates that an envelope did not contain any signatures.
+var ErrNoSignature = errors.New("no signature found")
 
 type EnvelopeVerifier struct {
 	providers []Verifier
